# api-contracts/p0/system-service.openapi.yaml
openapi: 3.1.0
info:
  title: Biel Life Camp System Service API
  version: 1.0.0
  description: P0 身份、组织、RBAC、第三方配置和操作审计契约。
servers:
  - url: /
...
paths:
...
  /api/system/v1/ehr-sync-runs:
    get:
      tags: [EhrSynchronization]
      operationId: listEhrSyncRuns
      summary: listEhrSyncRuns
      security:
        - ExternalBearer: []
      parameters:
        - $ref: './common.yaml#/components/parameters/PageSize'
      responses:
        '200':
          description: EHR 首次及每日全量人员同步批次。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EhrSyncRunPageApiResponse'
        '403':
          $ref: './common.yaml#/components/responses/Forbidden'
    post:
      tags: [EhrSynchronization]
      operationId: startEhrSyncRun
      summary: startEhrSyncRun
      security:
        - ExternalBearer: []
      parameters:
        - $ref: './common.yaml#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartEhrSyncRequest'
      responses:
        '202':
          description: 手工全量同步任务已执行；每日全量同步由 XXL-JOB 触发。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EhrSyncRunApiResponse'
        '403':
          $ref: './common.yaml#/components/responses/Forbidden'
        '409':
          $ref: './common.yaml#/components/responses/Conflict'
  /api/system/v1/ehr-sync-runs/{syncRunId}:
    get:
      tags: [EhrSynchronization]
      operationId: getEhrSyncRun
      summary: getEhrSyncRun
      security:
        - ExternalBearer: []
      parameters:
        - $ref: '#/components/parameters/SyncRunId'
      responses:
        '200':
          description: 全量同步统计和错误摘要。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EhrSyncRunApiResponse'
        '404':
          $ref: './common.yaml#/components/responses/NotFound'
...
components:
  securitySchemes:
    ExternalBearer:
      $ref: './common.yaml#/components/securitySchemes/ExternalBearer'
    InternalIdentity:
      $ref: './common.yaml#/components/securitySchemes/InternalIdentity'
    ServiceMtls:
      $ref: './common.yaml#/components/securitySchemes/ServiceMtls'
    GatewayServiceToken:
      type: apiKey
      in: header
      name: X-Gateway-Service-Token
  parameters:
...
    SyncRunId:
      name: syncRunId
      in: path
      required: true
      schema:
        $ref: './common.yaml#/components/schemas/DistributedId'
  schemas:
...
    EhrSyncRunApiResponse:
      allOf:
        - $ref: './common.yaml#/components/schemas/ApiResponse'
        - type: object
          properties:
            code:
              const: '0'
            errorMsg:
              const: ''
            data:
              $ref: '#/components/schemas/EhrSyncRun'
    EhrSyncRunPageApiResponse:
      allOf:
        - $ref: './common.yaml#/components/schemas/ApiResponse'
        - type: object
          properties:
            code:
              const: '0'
            errorMsg:
              const: ''
            data:
              $ref: '#/components/schemas/EhrSyncRunPage'
...
    EhrSyncRunType:
      type: string
      enum: [FULL_RECONCILIATION]
    EhrSyncRunStatus:
      type: string
      enum: [PENDING, RUNNING, SUCCEEDED, PARTIAL_SUCCEEDED, FAILED]
    EhrSyncRun:
      type: object
      additionalProperties: false
      required: [id, runType, triggerType, status, fetchedCount, insertedCount, updatedCount, resignedCount, roleInitializedCount, issueCount, createdAt]
      properties:
        id:
          $ref: './common.yaml#/components/schemas/DistributedId'
        runType:
          $ref: '#/components/schemas/EhrSyncRunType'
        triggerType:
          type: string
          enum: [DEPLOYMENT, SCHEDULED, MANUAL, RETRY]
        status:
          $ref: '#/components/schemas/EhrSyncRunStatus'
        fetchedCount:
          type: integer
          format: int64
          minimum: 0
        insertedCount:
          type: integer
          format: int64
          minimum: 0
        updatedCount:
          type: integer
          format: int64
          minimum: 0
        resignedCount:
          type: integer
          format: int64
          minimum: 0
        roleInitializedCount:
          type: integer
          format: int64
          minimum: 0
        issueCount:
          type: integer
          format: int64
          minimum: 0
        failureCode:
          type: [string, 'null']
          maxLength: 64
        failureDigest:
          type: [string, 'null']
          maxLength: 500
        startedAt:
          oneOf:
            - $ref: './common.yaml#/components/schemas/UtcDateTime'
            - type: 'null'
        completedAt:
          oneOf:
            - $ref: './common.yaml#/components/schemas/UtcDateTime'
            - type: 'null'
        createdAt:
          $ref: './common.yaml#/components/schemas/UtcDateTime'
    EhrSyncRunPage:
      type: object
      additionalProperties: false
      required: [items, page]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/EhrSyncRun'
        page:
          $ref: './common.yaml#/components/schemas/CursorPage'
    StartEhrSyncRequest:
      type: object
      additionalProperties: false
      required: [runType, reason, confirmationToken]
      properties:
        runType:
          $ref: '#/components/schemas/EhrSyncRunType'
        reason:
          type: string
          minLength: 5
          maxLength: 500
        confirmationToken:
          type: string
          minLength: 16
          maxLength: 512
...